Update launch_support.py delete_torch_dependencies() uses is_win32_standalone_build

+++ launch_support.py
@@ -147,17 +147,18 @@ def dependency_resolver():
     return dependencies
 
 def delete_torch_dependencies():
-    library_path = os.path.abspath(f'../python_embedded/Lib/site-packages')
-    file_paths = ['torch', 'torchaudio', 'torchvision', 'xformers',\
-        'pytorch_lightning', 'lightning-fabric']
-    for file_path in file_paths:
-        scratch_path = file_path
-        if os.path.exists(f'{library_path}/{file_path}'):
-            print(file_path)
-            shutil.rmtree(f'{library_path}/{file_path}', ignore_errors=True)
-        for scratch_path in glob.glob(f'{library_path}/{file_path}-*-info'):
-            print(scratch_path)
-            shutil.rmtree(scratch_path, ignore_errors=True)
+    if is_win32_standalone_build:
+        library_path = os.path.abspath(f'../python_embedded/Lib/site-packages')
+        file_paths = ['torch', 'torchaudio', 'torchvision', 'xformers',\
+            'pytorch_lightning', 'lightning-fabric']
+        for file_path in file_paths:
+            scratch_path = file_path
+            if os.path.exists(f'{library_path}/{file_path}'):
+                print(file_path)
+                shutil.rmtree(f'{library_path}/{file_path}', ignore_errors=True)
+            for scratch_path in glob.glob(f'{library_path}/{file_path}-*-info'):
+                print(scratch_path)
+                shutil.rmtree(scratch_path, ignore_errors=True)
     return
 
 def get_split_value(full_string):
